added remove

+++ src/stores/loans.js
@@ -5,32 +5,49 @@ export const useLoanStore = defineStore('loan', () => {
 
 
     const savedLendersJSON = localStorage.getItem('lenders');
-    const  lenders =ref(JSON.parse(savedLendersJSON)?? [])
+    const lenders = ref(JSON.parse(savedLendersJSON) ?? [])
 
-    const addLender= (lender, closure)=>{
-          lender.id=  generateId();
-          lenders.value.push(lender)
-          localStorage.setItem('lenders', JSON.stringify(lenders.value));
-          console.log(lenders.value)
+    const addLender = (lender, closure) => {
+        lender.id = generateId();
+        lender.total_credit = parseInt(lender.total_credit);
+        lenders.value.push(lender)
+        localStorage.setItem('lenders', JSON.stringify(lenders.value));
+        console.log(lenders.value)
 
-         closure()
+        closure()
     }
 
+    const removeLender = (id) => {
+        if (confirm('Are you sure you want to remove this lender?')) {
+            console.log(id)
+            const index = lenders.value.findIndex(lender => lender.id === id);
+
+
+            if (index !== -1) {
+                lenders.value.splice(index, 1)
+                localStorage.setItem('lenders', JSON.stringify(lenders.value));
+            }
+        }
+    };
+
+
     const generateId = () => {
-       
+
+
         if (lenders.value.length === 0) {
             return 1;
         }
         const lastItemId = lenders.value[lenders.value.length - 1].id;
-    
+
         return lastItemId + 1;
+
     };
 
-    const addCredit=(lender,items)=>{
+    const addCredit = (lender, items) => {
 
         console.log(items)
     }
-    
-  
-    return { lenders,addLender ,addCredit}
-  })
+
+
+    return { lenders, addLender, addCredit, removeLender }
+})
